task  5

+++ 0x06-pointers_arrays_strings/5-string_toupper.c
@@ -11,7 +11,7 @@ char *string_toupper(char *a)
 {
 	int i;
 
-	for (i = 0; i < strlen(a) ; i++)
+	for (i = 0; i < (int)strlen(a) ; i++)
 	{
 		if (a[i] >= 'a' && a[i] <= 'z')
 			a[i] = a[i] - 32;
